Extracted a ResultFormatter interface and cleaned up results
Extracted the ResultFormat interface so that Results are not bound to a String representation or hardcoded formats. With this the Results get a lot cleaner and concerns a better seperated. Also the Result interfaces and implementations got cleaned up quite a bit and more was moved to defaults instead of implementation. Implemented a DefaultResultFormatter and moved the ResourceBundleMessageResolver to the core where it belongs and added a class for instantiating defaults.